+++ src/main/java/com/a5lab/axion/domain/radar/RadarServiceImpl.java
@@ -85,14 +85,14 @@ public class RadarServiceImpl implements RadarService {
       // Find another primary radar
       List<Radar> radarList = radarRepository.findByPrimary(true);
       for (Radar radarItem : radarList) {
-        modelErrorList.addAll(new RadarPrimaryApprover(messageSource, radarDto, radarItem).approve());
+        modelErrorList.addAll(new RadarPrimarySaveApprover(messageSource, radarDto, radarItem).approve());
       }
     }
 
     // Check uniqueness by title
     List<Radar> radarList = this.radarRepository.findByTitle(radarDto.getTitle());
     for (Radar radarItem : radarList) {
-      modelErrorList.addAll(new RadarUniqueTitleApprover(messageSource, radarDto, radarItem).approve());
+      modelErrorList.addAll(new RadarUniqueTitleSaveApprover(messageSource, radarDto, radarItem).approve());
     }
 
     // Throw exception if violations are exists
