+++ src/main/java/com/a5lab/axion/domain/segment/RadarActiveDeleteApprover.java
@@ -1,4 +1,4 @@
-package com.a5lab.axion.domain.ring;
+package com.a5lab.axion.domain.segment;
 
 
 import java.util.LinkedList;
@@ -13,17 +13,17 @@ import com.a5lab.axion.domain.ModelError;
 import com.a5lab.axion.domain.ValidationException;
 
 @RequiredArgsConstructor
-public class RadarActiveApprover implements ModelApprover {
+public class RadarActiveDeleteApprover implements ModelApprover {
 
   private final MessageSource messageSource;
 
-  private final Ring ring;
+  private final Segment segment;
 
   @Override
   public List<ModelError> approve() throws ValidationException {
-    if (ring.getRadar().isActive()) {
+    if (segment.getRadar().isActive()) {
       return List.of(new ModelError("unable_to_delete_active_radar",
-          messageSource.getMessage("ring.error.unable_to_delete_active_radar", null,
+          messageSource.getMessage("segment.error.unable_to_delete_active_radar", null,
               LocaleContextHolder.getLocale()), null));
     }
     return new LinkedList<>();
